Backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
pillow